warn only once for unmaintained plugins
git-svn-id: http://josm.openstreetmap.de/svn/trunk@2261 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/plugins/PluginHandler.java
@@ -75,11 +75,14 @@ public class PluginHandler {
                 );
             }
         }
-        for (String p : unmaintained) {
-            if (plugins.contains(p) && disablePlugin(tr("<html>Loading of {0} plugin was requested."
-            +"<br>This plugin is no longer developed and very likely will produce errors."
-            +"<br>It should be disabled.<br>Delete from preferences?</html>", p), p))
-                plugins.remove(p);
+        if(early)
+        {
+            for (String p : unmaintained) {
+                if (plugins.contains(p) && disablePlugin(tr("<html>Loading of {0} plugin was requested."
+                +"<br>This plugin is no longer developed and very likely will produce errors."
+                +"<br>It should be disabled.<br>Delete from preferences?</html>", p), p))
+                    plugins.remove(p);
+            }
         }
 
         if (plugins.isEmpty())
